chore: cover component definition endpoints in pipeline API tests
