+++ integration-test/pipeline/rest.js
@@ -9,7 +9,9 @@ import {
   pipelinePublicHost,
 } from "./const.js";
 
+import * as connectorDefinition from "./rest-connector-definition.js";
 import * as constant from "./const.js";
+import * as operatorDefinition from "./rest-operator-definition.js";
 import * as pipelinePublic from './rest-pipeline-public.js';
 import * as pipelinePublicWithJwt from './rest-pipeline-public-with-jwt.js';
 import * as pipelinePrivate from './rest-pipeline-private.js';
@@ -126,6 +128,10 @@ export default function (data) {
     trigger.CheckTrigger(data)
     triggerAsync.CheckTrigger(data)
 
+    connectorDefinition.CheckList(data)
+    connectorDefinition.CheckGet(data)
+    operatorDefinition.CheckList()
+    operatorDefinition.CheckGet()
   }
 }
